@@ -4,10 +4,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "node-sass": "^4.12.0",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "react-scripts": "3.1.1"
+    "node-sass": "^9.0.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-scripts": "5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -32,7 +32,7 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^2.1.1",
-    "react-router-dom": "^5.0.1"
+    "gh-pages": "^6.1.1",
+    "react-router-dom": "^6.23.1"
   }
 }
