@@ -1,6 +1,6 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import { BrowserRouter }  from 'react-router-dom';
+import { createRoot } from 'react-dom/client';
 import './sass/main.scss';
 import App from './containers/App';
 import * as serviceWorker from './serviceWorker';
@@ -11,7 +11,9 @@ const app = (
     </BrowserRouter>
 );
 
-ReactDOM.render(app, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container); 
+root.render(app);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
